feat: ✨️添加 JSON 控制台过滤器并优化编辑器推送逻辑     - 新增 JsonConsoleFilter 类实现 JSON 日志的识别和处理     - 添加 JsonFilterProvider 类提供默认过滤器     - 重构 JsonHelperAction 类，使用新的 JsonEditorPushProvider 推送内容到编辑器     - 新增 JsonEditorPushProvider 类统一处理编辑器推送逻辑     - 更新插件配置，注册新的过滤器提供者 build: 🔨️️️️版本提升 0.7.11 -> 0.7.11.2
