+++ build.gradle
@@ -7,7 +7,7 @@ plugins {
 }
 
 group 'com.acme'
-version '0.7.11'
+version '0.7.11.2'
 
 repositories {
     mavenCentral()
