+++ src/main/java/com/acme/json/helper/core/parser/JsonParser.java
@@ -6,7 +6,7 @@ import com.acme.json.helper.core.parser.converter.*;
 import java.util.List;
 
 /**
- * JSON转任何
+ * JSON转任何文件
  * @see AnyFile
  * @author 拒绝者
  * @date 2025-01-26
